(pure-move) merege tensor_apis.hpp with tensor_utils.hpp
