+++ tt_metal/impl/sources.cmake
@@ -120,7 +120,6 @@ set(IMPL_SRC
     ${CMAKE_CURRENT_SOURCE_DIR}/tensor/topology/tensor_topology.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tensor/mesh_tensor.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tensor/host_tensor.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/tensor/tensor_utils.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tensor/tensor_impl.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tensor/tensor_apis.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tensor/host_tensor_factory.cpp
